Moved appointments components.
