+++ src/applications/check-in/components/AppointmentDisplay/AppointmentAction.jsx
@@ -1,14 +1,14 @@
 import React, { useCallback, useState } from 'react';
 import PropTypes from 'prop-types';
 import { useDispatch } from 'react-redux';
-import { api } from '../../../api';
+import { api } from '../../api';
 
-import { useFormRouting } from '../../../hooks/useFormRouting';
-import { ELIGIBILITY, areEqual } from '../../../utils/appointment/eligibility';
+import { useFormRouting } from '../../hooks/useFormRouting';
+import { ELIGIBILITY, areEqual } from '../../utils/appointment/eligibility';
 import recordEvent from 'platform/monitoring/record-event';
 import format from 'date-fns/format';
 
-import { appointmentWasCheckedInto } from '../../../actions/day-of';
+import { appointmentWasCheckedInto } from '../../actions/day-of';
 
 const AppointmentAction = props => {
   const { appointment, router, token } = props;
