+++ src/applications/check-in/day-of/pages/Confirmation/MultipleAppointments.jsx
@@ -8,7 +8,7 @@ import { focusElement } from 'platform/utilities/ui';
 import BackToHome from '../../components/BackToHome';
 import BackToAppointments from '../../../components/BackToAppointments';
 import Footer from '../../components/Footer';
-import AppointmentLocation from '../../components/AppointmentDisplay/AppointmentLocation';
+import AppointmentLocation from '../../../components/AppointmentDisplay/AppointmentLocation';
 import TravelPayReimbursementLink from '../../../components/TravelPayReimbursementLink';
 
 const MultipleAppointments = props => {
